Update the Android app
Target API 34. Remove lots of extra code, refactor parts of the app to simplify the logic. Update the UI to support both landscape and portrait orientations and add a proper loading skeleton. Update to latest Kotlin, Compose and Jetpack libraries. Remove extra dependencies. Remove tests as we're not using them.